Factor out ViewerBase and ViewerElementBase
