+++ packages/tools/viewer/src/index.ts
@@ -1,20 +1,10 @@
-export type {
-    CameraAutoOrbit,
-    EnvironmentOptions,
-    HotSpot,
-    LoadModelOptions,
-    Model,
-    PostProcessing,
-    ShadowQuality,
-    ToneMapping,
-    ViewerDetails,
-    ViewerHotSpotQuery,
-    ViewerOptions,
-} from "./viewer";
+export type { CameraAutoOrbit, EnvironmentOptions, HotSpot, PostProcessing, ShadowQuality, ToneMapping, ViewerHotSpotQuery } from "./viewerBase";
+export type { LoadModelOptions, Model, ViewerDetails, ViewerOptions } from "./viewer";
 export type { CanvasViewerOptions } from "./viewerFactory";
-export type { ViewerElementEventMap } from "./viewerElement";
+export type { ViewerElementEventMap } from "./viewerElementBase";
 
-export { CreateHotSpotFromCamera, DefaultViewerOptions, Viewer, ViewerHotSpotResult } from "./viewer";
+export { ViewerHotSpotResult } from "./viewerBase";
+export { CreateHotSpotFromCamera, DefaultViewerOptions, Viewer } from "./viewer";
 export { ConfigureCustomViewerElement, HTML3DElement, ViewerElement } from "./viewerElement";
 export { CreateViewerForCanvas } from "./viewerFactory";
 export { HTML3DAnnotationElement } from "./viewerAnnotationElement";
